Don't send duplicate messages.
We have to switch to using watch because watchCollection because of https://github.com/angular/angular.js/issues/2621

Note - this does not fix bug 15, - there is either a bug in django-websocket-redis or we are using it incorrectly.   (if you edit a field quickly, then it will continually send us the last few edits).